Added anims for checking the menu and getting ready to battle.
The check menu anim should loop image 11 and 12 to make the menu from
the orb "flicker".

The battle anim should loop from image 5 to image 9 to keep the cycle
steady.

The looping part of the battle ready anim is also an example of the idle
stance we will animate in code for inside the battles. Basically im
gonna cut out the arms and then give you the arms and the body idle
seperately. Heck lemme do that right now!